move to defs/ for autoloading
