+++ hooli_data_eng/defs/custom_ingest/assets.py
@@ -2,7 +2,7 @@ from datetime import timedelta
 import dagster as dg
 import pandas as pd
 
-from hooli_data_eng.custom_ingest.resources import RawDataAPI
+from hooli_data_eng.defs.custom_ingest.resources import RawDataAPI
 from hooli_data_eng.utils.kind_helpers import get_kind
