+++ hooli_data_eng/defs/kubernetes/assets.py
