+++ hooli_data_eng/defs/kubernetes/resources.py
