+++ hooli_data_eng/defs/spark/definitions.py
@@ -1,5 +1,5 @@
 import dagster as dg
-from hooli_data_eng.spark.resources import resource_def
+from hooli_data_eng.defs.spark.resources import resource_def
 from hooli_data_eng.utils import get_env
 
 defs = dg.Definitions(
